+++ plugins/events-correlation-engine/src/test/java/org/opensearch/plugin/correlation/EventsCorrelationPluginTests.java
@@ -1,0 +1,19 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+package org.opensearch.plugin.correlation;
+
+import org.junit.Assert;
+import org.opensearch.test.OpenSearchTestCase;
+
+public class EventsCorrelationPluginTests extends OpenSearchTestCase {
+
+    public void testDummy() {
+        Assert.assertEquals(1, 1);
+    }
+}